requirements.txt
check-jsonschema==0.37.1
pre-commit==4.5.1
pytest==9.0.3
pytest-cov==7.1.0
pytest-timeout==2.4.0
PyYAML==6.0.3
ruff==0.15.10
pandas==3.0.2
tabulate==0.10.0